// src/types/partials/SpecimenContainer.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TQuantity } from '../partials/Quantity';
import { TReference } from '../partials/Reference';

export type TSpecimenContainer = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    identifier?: TIdentifier[];
    description?: String;
    type?: TCodeableConcept;
    capacity?: TQuantity;
    specimenQuantity?: TQuantity;
    additiveCodeableConcept?: TCodeableConcept;
    additiveReference?: TReference;
};